requirements.txt
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
cachetools==4.2.0
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
importlib-metadata==3.1.1
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
packaging==20.7
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pipenv==2018.11.26
polygon-api-client==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyarrow==2.0.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.3
pylint==2.4.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-chess==0.28.3
python-dateutil==2.8.1
pytz==2020.4
pyzmq==20.0.0
requests==2.25.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
smmap==3.0.4
streamlit==0.72.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.0
tzlocal==2.1
urllib3==1.25.8
validators==0.18.1
virtualenv==16.7.5
virtualenv-clone==0.5.3
watchdog==1.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.4.0
